Update Crop Non-destructive functionality
On cropping the original image get's copied as a file and as an asset so you can work with the cropped version in Craft. This is instead of copying the original file into a folder you can only access through ftp. Now you can access the original and cropped version through the assets in Craft.