Use ternary operator instead of null coalese operator
I was noticing that I couldn't upload any files anymore since the 3.0.5 release because the $imageWidth and $imageHeight var returned an empty string.
Looking into the getSettingForAssetSource function it seems like when using the coalese operator the first part is always true, even if empty (because it's always defined). The second part, the general settings are never reached.
When using the ternary operator instead things seem to work correctly.